Remove obsolete Managed Lines during migration
