+++ changelog.txt
@@ -5,7 +5,7 @@ Date: 2026-07-25
     - Added character-only bootstrap recipes for hand-grown pentapod eggs on Gleba and hand-grown biter eggs on Nauvis.
   Changes:
     - Replaced Nauvis biter egg ingress and bioflux egress Managed Lines with Nauvis-only hand-crafted biter eggs.
-    - Existing saves stash configured biter egg and bioflux Managed Lines and refund placed lines as generic Managed Line items.
+    - Existing saves remove configured biter egg and bioflux Managed Lines; placed obsolete lines are not refunded.
   Bugfixes:
     - Base screenshots now capture the supported planet currently being viewed, use its current Square size, and include its name in the filename.
 ---------------------------------------------------------------------------------------------------
